AutuhTest/AuthUSer/requirements.txt
# Django-related dependencies
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# MongoDB-related dependencies
djongo==1.3.6
dnspython==2.6.1
pymongo==3.12.1

# Other dependencies
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.2.4
typing-extensions==4.10.0
tzdata==2024.1